FAT-58349: decompose T3NewsListFragment and some fixes
