+++ app/src/main/java/com/example/pavliukovandersen/T3NewsListFragment.kt
@@ -1,7 +1,6 @@
 package com.example.pavliukovandersen
 
 import android.os.Bundle
-import android.util.Log
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
@@ -13,15 +12,13 @@ import androidx.appcompat.widget.Toolbar
 import androidx.lifecycle.lifecycleScope
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.swiperefreshlayout.widget.SwipeRefreshLayout
+import com.example.pavliukovandersen.objects.CryptoUtil
 import com.example.pavliukovandersen.databinding.T3FragmentNewsListBinding
+import com.example.pavliukovandersen.objects.DateUtil
 import com.example.pavliukovandersen.retrofit.ArticleDto
 import com.example.pavliukovandersen.retrofit.RetrofitInstance
 import com.example.pavliukovandersen.retrofit.SourceDto
 import kotlinx.coroutines.launch
-import javax.crypto.Cipher
-import javax.crypto.spec.SecretKeySpec
-import java.util.Base64
-import java.util.Calendar
 
 class T3NewsListFragment : Fragment() {
     private lateinit var toolbar: Toolbar
@@ -30,11 +27,8 @@ class T3NewsListFragment : Fragment() {
     private lateinit var newsAdapter: T3NewsAdapter
     private lateinit var swipeRefreshLayout: SwipeRefreshLayout
 
-    private val apiKey: String get() = getKey()
-
-    @get:JvmName("retrieveCurrentDate")
-    val currentDate: String
-        get() = getCurrentDate()
+    private val apiKey: String get() = CryptoUtil.getDecryptedKey()
+    private val currentDate: String get() = DateUtil.getCurrentDate()
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?
@@ -110,16 +104,5 @@ class T3NewsListFragment : Fragment() {
             .replace(R.id.displayed_screen_fl, new)
             .addToBackStack(null)
             .commit()
-    }
-
-    private fun getKey(): String {
-        val cipher = Cipher.getInstance("AES/ECB/PKCS5Padding")
-        cipher.init(Cipher.DECRYPT_MODE, SecretKeySpec("0123456789abcdef".toByteArray(), "AES"))
-        return String(cipher.doFinal(Base64.getDecoder().decode(Constants.KEY)))
-    }
-
-    private fun getCurrentDate(): String {
-        val c = Calendar.getInstance()
-        return "${c.get(Calendar.YEAR)}-${c.get(Calendar.MONTH) + 1}-${c.get(Calendar.DAY_OF_MONTH)}"
     }
 }
